+++ server/tsuserver.py
@@ -56,8 +56,8 @@ class TsuserverDR:
         self.release = 4
         self.major_version = 3
         self.minor_version = 0
-        self.segment_version = 'b112'
-        self.internal_version = 'M210101g'
+        self.segment_version = 'b113'
+        self.internal_version = 'M210101h'
         version_string = self.get_version_string()
         self.software = 'TsuserverDR {}'.format(version_string)
         self.version = 'TsuserverDR {} ({})'.format(version_string, self.internal_version)
@@ -300,7 +300,7 @@ class TsuserverDR:
             Clients connected to the server.
 
         """
-        return sorted(self.client_manager.clients, key=lambda c: c.id)
+        return sorted(self.client_manager.clients)
 
     def get_player_count(self):
         # Ignore players in the server selection screen.